requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==4.0.0
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.7
beautifulsoup4==4.10.0
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
channels-redis==3.3.1
chardet==3.0.4
charset-normalizer==2.0.4
click==8.0.3
constantly==15.1.0
crispy-tailwind==0.5.0
cryptography==3.4.8
daphne==3.0.2
deep-translator==1.5.4
defusedxml==0.7.1
Django==3.2.6
django-allauth==0.45.0
django-appconf==1.0.4
django-ckeditor==6.1.0
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.2
django-environ==0.5.0
django-js-asset==1.2.2
django-redis==5.0.0
django-select2==7.7.1
django-widget-tweaks==1.4.8
h11==0.9.0
h2==3.2.0
hiredis==2.0.0
hpack==3.0.0
hstspreload==2021.11.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
libretranslatepy==2.1.1
lxml==4.6.4
msgpack==1.0.2
oauthlib==3.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
pyOpenSSL==21.0.0
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
rfc3986==1.5.0
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3
sqlparse==0.4.1
toml==0.10.2
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.2
urllib3==1.26.6
zope.interface==5.4.0
